Move params to cli module
