+++ org.hl7.fhir.validation.cli/src/main/java/org/hl7/fhir/validation/cli/tasks/AiTestsTask.java
@@ -6,7 +6,7 @@ import org.hl7.fhir.utilities.SystemExitManager;
 import org.hl7.fhir.utilities.TimeTracker;
 import org.hl7.fhir.validation.ai.AITests;
 import org.hl7.fhir.validation.service.model.ValidationContext;
-import org.hl7.fhir.validation.service.utils.Params;
+import org.hl7.fhir.validation.cli.param.Params;
 
 public class AiTestsTask extends StandaloneTask{
   @Override
