+++ org.hl7.fhir.validation.cli/src/main/java/org/hl7/fhir/validation/cli/tasks/PreloadCacheTask.java
@@ -3,7 +3,7 @@ package org.hl7.fhir.validation.cli.tasks;
 import org.hl7.fhir.convertors.analytics.PackageVisitor;
 import org.hl7.fhir.utilities.TimeTracker;
 import org.hl7.fhir.validation.service.model.ValidationContext;
-import org.hl7.fhir.validation.service.utils.Params;
+import org.hl7.fhir.validation.cli.param.Params;
 import org.hl7.fhir.validation.packages.PackageCacheDownloader;
 
 import java.io.PrintStream;
