+++ org.hl7.fhir.validation.cli/src/main/java/org/hl7/fhir/validation/cli/tasks/SpecialTask.java
@@ -5,7 +5,7 @@ import java.io.PrintStream;
 import org.hl7.fhir.utilities.TimeTracker;
 import org.hl7.fhir.utilities.filesystem.ManagedFileAccess;
 import org.hl7.fhir.validation.service.model.ValidationContext;
-import org.hl7.fhir.validation.service.utils.Params;
+import org.hl7.fhir.validation.cli.param.Params;
 import org.hl7.fhir.validation.special.R4R5MapTester;
 
 public class SpecialTask extends StandaloneTask{
