+++ org.hl7.fhir.validation.cli/src/main/java/org/hl7/fhir/validation/cli/tasks/TestsTask.java
@@ -5,7 +5,7 @@ import java.io.PrintStream;
 import org.hl7.fhir.utilities.TimeTracker;
 import org.hl7.fhir.validation.service.model.ValidationContext;
 import org.hl7.fhir.validation.service.utils.Display;
-import org.hl7.fhir.validation.service.utils.Params;
+import org.hl7.fhir.validation.cli.param.Params;
 import org.hl7.fhir.validation.testexecutor.TestExecutor;
 import org.hl7.fhir.validation.testexecutor.TestExecutorParams;
