+++ org.hl7.fhir.validation.cli/src/main/java/org/hl7/fhir/validation/cli/tasks/TxTestsTask.java
@@ -12,7 +12,7 @@ import org.hl7.fhir.utilities.json.JsonException;
 import org.hl7.fhir.utilities.json.model.JsonObject;
 import org.hl7.fhir.utilities.json.parser.JsonParser;
 import org.hl7.fhir.validation.service.model.ValidationContext;
-import org.hl7.fhir.validation.service.utils.Params;
+import org.hl7.fhir.validation.cli.param.Params;
 import org.hl7.fhir.validation.special.TxTester;
 
 public class TxTestsTask extends StandaloneTask{
